Add files via upload
My project is designed to facilitate user registration by providing a registration form. Upon completion of this form, users create an account. Subsequently, upon successful account creation, they are directed to the login page. After entering the necessary credentials on the login page, users are then redirected to the homepage of the application.